Fix all code review findings from Nextcloud + certificate feature
- Move file I/O (photo and certificate copy) off main thread in AddEditFragment
  using suspend funs with withContext(Dispatchers.IO)
- Call conn.disconnect() after each HTTP response in NextcloudClient
- Replace full-table-scan in ArtworkDetailViewModel with a direct
  getArtworkFlow(id) DAO query for O(1) lookup
- Reset savedId StateFlow immediately after consumption to prevent
  re-navigation when AddEdit returns to STARTED from the back stack
- Fix DashboardViewModel.computeValueState to return Unavailable
  (not Loading) when no priced artworks exist
- Add NextcloudClient.Result.Transient for network/timeout errors;
  NextcloudBackupWorker now returns Result.retry() for transient failures
- Migrate NextcloudPreferences to EncryptedSharedPreferences with
  AES256_GCM master key and backup exclusion rules
- Rewrite BackupExporter to include only files referenced by artwork
  records (not the entire artworks/ directory) and add UTC timezone
- Add ZIP slip protection and UTC timezone to BackupImporter
- Add SSL warning dialog before enabling trust-all-certs in Nextcloud
- Change nav artworkId arg type from integer to long throughout
- Enable minifyEnabled true for release builds
- Add certificate PDF attachment feature with FileProvider open support

+++ app/src/main/java/com/example/artworksmanager/ui/collection/CollectionFragment.kt
@@ -34,7 +34,7 @@ class CollectionFragment : Fragment() {
 
     private val adapter = ArtworkAdapter { artwork ->
         findNavController().navigate(
-            CollectionFragmentDirections.actionCollectionToDetail(artwork.id.toInt())
+            CollectionFragmentDirections.actionCollectionToDetail(artwork.id)
         )
     }
