+++ app/src/main/java/com/example/artworksmanager/ui/dashboard/DashboardFragment.kt
@@ -42,7 +42,7 @@ class DashboardFragment : Fragment() {
 
     private val recentAdapter = RecentArtworkAdapter { artwork ->
         findNavController().navigate(
-            DashboardFragmentDirections.actionDashboardToDetail(artwork.id.toInt())
+            DashboardFragmentDirections.actionDashboardToDetail(artwork.id)
         )
     }
